Update websockets version range in requirements.txt
Specified the version range for websockets in `requirements.txt` to be
">=10, <14" to ensure compatibility with the current CSMS implementation.
I encountered issues with websockets version 14 and, while unable to fix
the code for that version, limiting the version solves the problem.

Additionally, I removed imports of websockets.asyncio, as they were not
being used and restricted the websockets version to 13.

Signed-off-by: Akira - Cleber Akira Nakandakare <cakira@cpqd.com.br>